In model classes, use only identity fields in hashCode().
In most cases, this means their ID number field; in Loan,
the book, borrower, and branch. This prevents memory leaks,
as hashing based on fields that can be changed by setters
(or themselves have hashCode() methods that could change)
causes HashSets and HashMaps to lose the hashed object
in the old hash bucket.